wiki support alpha version

+++ src/wiki/TXML.cpp
@@ -65,9 +65,9 @@ string TXML::get_string ()
 	{
     string ret ;
     ret = "<" + name ;
-    for ( int a = 0 ; a < key.size() ; a++ )
+    for ( size_t a = 0 ; a < key.size() ; a++ )
     	{
-	    for ( int b = 0 ; b < key[a].length() ; b++ )
+	    for ( size_t b = 0 ; b < key[a].length() ; b++ )
 	    	{
       		if ( key[a][b] == ' ' ) key[a][b] = '_' ;
       		}
